requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
crispy-tailwind==1.0.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.7
django-crispy-forms==2.4
executing==2.2.1
flatbuffers==25.9.23
fonttools==4.60.1
fpdf2==2.8.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
jiter==0.11.1
kiwisolver==1.4.9
matplotlib==3.10.7
matplotlib-inline==0.2.1
mediapipe==0.10.21
ml_dtypes==0.5.3
numpy==1.26.4
openai==2.7.1
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
parso==0.8.5
pillow==12.0.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psycopg2==2.9.11
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scipy==1.16.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
